src/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
alembic==1.14.1
psycopg[binary]==3.2.3
pydantic==2.10.4
pydantic-settings==2.7.0
python-dotenv==1.0.1
httpx==0.28.1
tenacity==9.0.0
qdrant-client==1.12.2
pypdf==5.1.0
pdfminer.six==20231228
beautifulsoup4==4.12.3
readability-lxml==0.8.1
rank-bm25==0.2.2
sentence-transformers==3.3.1
openai==1.58.1
streamlit==1.41.1
requests==2.32.3
python-dateutil==2.9.0.post0
orjson==3.10.12
pytest==8.3.4
pytest-asyncio==0.25.0
